+++ client/src/components/ExpenseTracker.js
@@ -2,7 +2,7 @@ import { useEffect, useState } from "react";
 import axios from "axios";
 import "../styles/ExpenseTracker.css";
 import "../index.css";
-import { addDoc, query, where, onSnapshot, orderBy, updateDoc, getDocs, deleteDoc} from 'firebase/firestore';
+import { query, where, onSnapshot, orderBy, getDocs, deleteDoc} from 'firebase/firestore';
 import { transactionsCollection } from '../firebaseConfig';
 import { ToastContainer, toast } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
@@ -16,8 +16,7 @@ import { ReactComponent as Others } from '../icons/others.svg';
 import {ReactComponent as Plus} from '../icons/plus1.svg';
 import DoughnutChart from "./DoughnutChart";
 import MyLoader from "./TransactionLoading";
-
-
+import TransactionForm from "./TransactionForm";
 
 function ExpenseTracker() 
 {
@@ -28,14 +27,15 @@ function ExpenseTracker()
     const [transactionFilter, setTransactionFilter] = useState([]);
     const [transactionType,setTransactionType]=useState('')
     const [formData, setFormData] = useState({
-        email: "",
-        transactionType: "",
-        category: "",
-        date: "",
-        amount: "",
-        description: "",
-        transactionId: ""
+      email: "",
+      transactionType: "",
+      category: "",
+      date: "",
+      amount: "",
+      description: "",
+      transactionId: "",
     });
+    
     const [balance, setBalance] = useState(0);
     const [incoming, setIncoming] = useState(0);
     const [outgoing, setOutgoing] = useState(0);
@@ -73,18 +73,6 @@ function ExpenseTracker()
     , [userToken]);
 
 
-    function generateTransactionId()
-    {
-        const characters = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
-        let transactionId = '';
-        for (let i = 0; i < 10; i++) 
-        {
-            const randomIndex = Math.floor(Math.random() * characters.length);
-            transactionId += characters.charAt(randomIndex);
-        }
-        return transactionId;
-    }
-
     useEffect(() => 
     {
         let incoming = 0;
@@ -104,12 +92,6 @@ function ExpenseTracker()
         setOutgoing(outgoing);
         setBalance(incoming - outgoing);
     }, [transactions]);
-
-    const handleChange = (e) => 
-    {
-        const targetValue = e.target.name === 'amount' ? parseFloat(e.target.value) : e.target.value;
-        setFormData({ ...formData, [e.target.name]: targetValue });
-    };
 
     useEffect(() => 
     {
@@ -148,112 +130,6 @@ function ExpenseTracker()
         localStorage.removeItem('dateFilter')
       
     },[])
-    const handleSubmit = async (e) => 
-    {
-        e.preventDefault();
-        if (formData.transactionType === "")
-        {
-            toast.error("Please select a transaction type.");
-            return;
-        }
-        if (formData.transactionType === "Income" && formData.category !== "NULL")
-        {
-            setFormData({ ...formData, category: "NULL" });
-        }
-        if (formData.transactionType === "Expense" && formData.category === "")
-        {
-            toast.error("Please select a category.");
-            return;
-        }
-        if (formData.date === "")
-        {
-            toast.error("Please select a date.");
-            return;
-        }
-        if (formData.amount === "")
-        {
-            toast.error("Please enter an amount.");
-            return;
-        }
-        if (formData.description === "")
-        {
-            toast.error("Please enter a description.");
-            return;
-        }
-        if (editEnabled) 
-        {
-            console.log("edit clicked ", formData.transactionId);
-            try 
-            {
-                const querySnapshot = await getDocs(
-                  query(transactionsCollection, where('transactionId', '==', formData.transactionId))
-                );
-              
-                if (querySnapshot.empty) 
-                {
-                    toast.error('Transaction not found.');
-                } 
-                else 
-                {
-                    const transactionDoc = querySnapshot.docs[0].ref;
-                    const transactionObject = 
-                    {
-                        transactionType: formData.transactionType,
-                        category: formData.category,
-                        date: formData.date,
-                        amount: formData.amount,
-                        description: formData.description,
-                    };
-                
-                    await updateDoc(transactionDoc, transactionObject);
-                    setFormData({
-                        transactionType: "",
-                        category: "",
-                        date: "",
-                        amount: "",
-                        description: ""
-                    });
-                    setEditEnabled(false);
-                    toast.success('Transaction edited successfully.');
-                }
-            } 
-            catch (error) 
-            {
-                toast.error('Error editing transaction.');
-            }
-        }
-        else 
-        {
-            try 
-            {
-                const transactionObject = 
-                {   
-                    email: user.email,
-                    transactionType: formData.transactionType,
-                    category: formData.category,
-                    date: formData.date,
-                    amount: formData.amount,
-                    description: formData.description,
-                    transactionId: generateTransactionId(),
-                    created_at: new Date(),
-                };
-                await addDoc(transactionsCollection, transactionObject);
-                // console.log('Transaction ID:', newTransactionRef.id);
-                setFormData({
-                    transactionType: "",
-                    category: "",
-                    date: "",
-                    amount: "",
-                    description: ""
-                });
-                toast.success("Transaction added successfully.");
-            } 
-            catch (error) 
-            {
-                toast.error('Error adding transaction.');
-            }
-        }
-    };
 
     const handleEdit = (transactionId, index) => 
     {
@@ -316,11 +192,7 @@ function ExpenseTracker()
        
 
     }
-    const newTransaction = () => {
-        setEditEnabled(false)
-        let addedTransaction={transactionType:'',category:'',date:'',description:'',amount:''}
-        setFormData(addedTransaction)
-    }
+
     const changeDateFilter = (e) => {
         if (e.target.value) {
             let transactionsfilter = transactions.filter(item => item.date == e.target.value)
@@ -352,109 +224,7 @@ function ExpenseTracker()
                 <h1>&#x20B9;{balance}</h1>
             </div>
             <div className="formTransactions_container flex justify-evenly">
-                <div className="form_container border-2 rounded p-8">
-                    <div className="form-title">
-                    {
-                        editEnabled
-                            ?
-                            <>
-                                <h4 className="font-bold text-lg mb-10 edittransaction">Edit transaction
-                                
-                                
-                                </h4>
-                                <button type="button" class="btn-new mb-8" onClick={()=>newTransaction()} >New</button>
-                            </>
-                         
-                            : <>
-                                <h4 className="font-bold text-lg mb-10">Add new transaction</h4>
-                                
-                            </>
-                         
-                    }
-                    </div>
-                   
-                 
-                    <form className="flex gap-4">
-                        <div className="flex flex-col gap-2">
-                            <legend>Transaction Type</legend>
-                            <label style={{ cursor: 'pointer' }}>
-                                <input
-                                    type="radio"
-                                    name="transactionType"
-                                    value="Income"
-                                    checked={formData.transactionType === "Income"}
-                                    onChange={handleChange}
-                                    required
-                                    className="cursor-pointer"
-                                />&nbsp;
-                                Income
-                            </label>
-                            <label style={{ cursor: 'pointer' }}>
-                                <input
-                                    type="radio"
-                                    name="transactionType"
-                                    value="Expense"
-                                    checked={formData.transactionType === "Expense"}
-                                    onChange={handleChange}
-                                    required
-                                    className="cursor-pointer"
-                                />&nbsp;
-                                Expense
-                            </label>
-                            <label htmlFor="category" className="mt-4">Category</label>
-                            <select
-                                id="category"
-                                name="category"
-                                value={formData.category}
-                                onChange={handleChange}
-                                required
-                                style={{ cursor: formData.transactionType === "Income" ? "auto" : "pointer" }}
-                                disabled={formData.transactionType === "Income"}
-                            >
-                                <option value="NULL">Choose a category</option>
-                                <option value="Food">Food</option>
-                                <option value="Travel">Travel</option>
-                                <option value="Shopping">Shopping</option>
-                                <option value="Bills">Bills</option>
-                                <option value="Others">Others</option>
-                            </select>
-                            <label htmlFor="date" className="mt-4">Date</label>
-                            <input
-                                type="date"
-                                name="date"
-                                value={formData.date}
-                                onChange={handleChange}
-                                placeholder="Date"
-                                required
-                                className="cursor-pointer"
-                            />
-                        </div>
-                        <div className="flex flex-col gap-2">
-                            <label htmlFor="amount">Amount</label>
-                            <input
-                                type="number"
-                                name="amount"
-                                value={formData.amount}
-                                onChange={handleChange}
-                                placeholder="Amount"
-                                required
-                                className="border-2 p-2"
-                            />
-                            <label htmlFor="description" className="mt-4">Description</label>
-                            <input
-                                type="text"
-                                name="description"
-                                value={formData.description}
-                                onChange={handleChange}
-                                placeholder="Description"
-                                required
-                                className="border-2 p-2"
-                                autoComplete="off"
-                            />
-                            <button type="submit" className=" mt-4 text-white hover:text-gray-500 hover:bg-white border-[#c465c9] p-2 border transition-all duration-500" onClick={handleSubmit}> {editEnabled ? "Edit Transaction" : "Add Transaction"} </button>
-                        </div>
-                    </form>
-                </div>
+                <TransactionForm editEnabled={editEnabled} setEditEnabled={setEditEnabled} user={user} formData={formData} setFormData={setFormData} />
             
 
                 <div className="WalletDetails_container flex flex-col">
@@ -464,7 +234,7 @@ function ExpenseTracker()
                     </div>
                     <div className="transaction-group row d-flex">
                       
-                        <div class="form-group col-md-4">
+                        <div className="form-group col-md-4">
                             <h4>Transactions</h4>
                             
                         <select id="inputState" class="form-control " onChange={(e)=>TransactionTypeChange(e)} value={transactionType}>
